更新了预售的服务层

+++ goods/src/main/java/cn/edu/xmu/goods/model/bo/PresaleActivity.java
@@ -9,13 +9,13 @@ import java.util.Map;
 @Data
 public class PresaleActivity {
     public enum PresaleStatus {
-        PENDING(0,"尚未开始"),
-        DEPOSIT(1,"支付定金"),
-        BALANCE(2,"支付尾款"),
-        FINISHED(3,"已结束"),
-        CANCELED(4,"已取消");
+        PENDING((byte)0,"尚未开始"),
+        DEPOSIT((byte)1,"支付定金"),
+        BALANCE((byte)2,"支付尾款"),
+        FINISHED((byte)3,"已结束"),
+        CANCELED((byte)4,"已取消");
 
-        private static final Map<Integer, PresaleStatus> typeMap;
+        private static final Map<Byte, PresaleStatus> typeMap;
         static { //由类加载机制，静态块初始加载对应的枚举属性到map中，而不用每次取属性时，遍历一次所有枚举值
             typeMap = new HashMap();
             for (PresaleStatus enum1 : values()) {
@@ -23,10 +23,10 @@ public class PresaleActivity {
             }
         }
 
-        private int code;
+        private byte code;
         private String description;
 
-        PresaleStatus(int code, String description) {
+        PresaleStatus(byte code, String description) {
             this.code = code;
             this.description = description;
         }
@@ -35,7 +35,7 @@ public class PresaleActivity {
             return typeMap.get(code);
         }
 
-        public Integer getCode() {
+        public Byte getCode() {
             return code;
         }
